rename header/index.jsx file to navbar.jsx
