+++ client/src/components/header/navbar.jsx
@@ -1,5 +1,44 @@
-import { Heading } from '@chakra-ui/react'
-export default function LoadHeader()
-{
-    return <Heading>Lol</Heading>;
-}
+import {
+    Box,
+    Button,
+    ButtonGroup,
+    Container,
+    Flex,
+    HStack,
+    IconButton,
+    useBreakpointValue,
+  } from '@chakra-ui/react'
+  import { FiMenu } from 'react-icons/fi'
+ // import { Logo } from 'client/src/components/header/Logo.jpg'
+  
+export default function App () {
+    const isDesktop = useBreakpointValue({ base: false, lg: true })
+    return (
+      <Box as="section" pb={{ base: '12', md: '24' }}>
+        <Box as="nav" bg="bg-surface" boxShadow="sm">
+          <Container py={{ base: '4', lg: '5' }}>
+            <HStack spacing="10" justify="space-between">
+              {isDesktop ? (
+                <Flex justify="space-between" flex="1">
+                  <ButtonGroup variant="link" spacing="8">
+                    {['Homepage', 'Products', 'About us', 'Contact', 'Checkout'].map((item) => (
+                      <Button key={item}>{item}</Button>
+                    ))}
+                  </ButtonGroup>
+                  <HStack spacing="3">
+                    <Button variant="ghost">Sign in</Button>
+                  </HStack>
+                </Flex>
+              ) : (
+                <IconButton
+                  variant="ghost"
+                  icon={<FiMenu fontSize="1.25rem" />}
+                  aria-label="Open Menu"
+                />
+              )}
+            </HStack>
+          </Container>
+        </Box>
+      </Box>
+    )
+  }
